Moved Components from SnabbleAssetProviding to Components
