+++ ScanAndGo/Shopping/Views/BarcodeSearchView.swift
@@ -9,6 +9,7 @@ import SwiftUI
 
 import SnabbleCore
 import SnabbleAssetProviding
+import SnabbleComponents
 import SnabbleUI
 
 struct BarcodeSearchRowView: View {
