+++ ScanAndGo/Shopping/Views/CheckoutView.swift
@@ -10,6 +10,7 @@ import Combine
 
 import SnabbleCore
 import SnabbleAssetProviding
+import SnabbleComponents
 import SnabbleUI
 
 struct CheckoutView: View {
