+++ User/Sources/Views/UserView.swift
@@ -6,7 +6,9 @@
 //
 
 import SwiftUI
+
 import SnabbleAssetProviding
+import SnabbleComponents
 
 private extension UserField {
     func next(in array: [UserField]) -> UserField? {
